+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to
 ## Fixed
 
 - 🐛(frontend) fix collaboration error #684
+- 🔒️(back) restrict access to favorite_list endpoint #690
 
 ## [2.3.0] - 2025-03-03